# TupleDescriptor performance improvements