--- a/Orm/Xtensive.Orm/Tuples/AccessorDelegates.cs
+++ b/Orm/Xtensive.Orm/Tuples/AccessorDelegates.cs
@@ -16,7 +16,7 @@ namespace Xtensive.Tuples
   /// <param name="descriptor">Field descriptor.</param>
   /// <param name="fieldState">State of a field.</param>
   /// <returns></returns>
-  internal delegate TValue GetValueDelegate<TValue>(PackedTuple tuple, PackedFieldDescriptor descriptor, out TupleFieldState fieldState);
+  internal delegate TValue GetValueDelegate<TValue>(PackedTuple tuple, ref PackedFieldDescriptor descriptor, out TupleFieldState fieldState);
 
   /// <summary>
   /// Incapsulates <see cref="Tuple.SetValue{T}"/> method.
@@ -25,5 +25,5 @@ namespace Xtensive.Tuples
   /// <param name="tuple">Tuple to use.</param>
   /// <param name="descriptor">Field descriptor.</param>
   /// <param name="value">A value.</param>
-  internal delegate void SetValueDelegate<TValue>(PackedTuple tuple, PackedFieldDescriptor descriptor, TValue value);
+  internal delegate void SetValueDelegate<TValue>(PackedTuple tuple, ref PackedFieldDescriptor descriptor, TValue value);
 }