--- a/Orm/Xtensive.Orm/Tuples/Packed/PackedFieldDescriptor.cs
+++ b/Orm/Xtensive.Orm/Tuples/Packed/PackedFieldDescriptor.cs
@@ -9,7 +9,7 @@ using System;
 namespace Xtensive.Tuples.Packed
 {
   [Serializable]
-  internal sealed class PackedFieldDescriptor
+  internal struct PackedFieldDescriptor
   {
     public FieldPackingType PackingType;