… Orm/Xtensive.Orm/Tuples/Packed/PackedFieldDescriptorComparer.cs
… Orm/Xtensive.Orm/Tuples/Packed/PackedTuple.cs
… Orm/Xtensive.Orm/Tuples/Packed/TupleLayout.cs
… Orm/Xtensive.Orm/Tuples/Tuple.cs
… Orm/Xtensive.Orm/Tuples/TupleDescriptor.cs
… Orm/Xtensive.Orm/Tuples/TupleExtensions.cs
… Orm/Xtensive.Orm/Xtensive.Orm.csproj