This PR is to improve performance of `TupleDescriptor.Create()` method.
Briefly, the following modifications where made
- `PackedFieldDescriptor` is not a `class` but `struct` now. It is more compact than it used to be plus it passed everywhere by reference. As a consequence memory load is reduced without any performance hit.
- No `Array.Sort()` method calls anymore. Instead we achieve correct field order by counting values of the same size.
- No calls to `StripNullable` extension method. Instead we use the fact underlying type is already known.

and all this led to considerable performance boost

```
Before:

|                 Method |      N |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
| TupleDescriptor_Create | 100000 | 5,317.3 ns | 34.32 ns | 32.10 ns | 0.8926 |     - |     - |    3760 B |
|         Tuple_GetValue | 100000 |   456.5 ns |  2.52 ns |  2.23 ns |      - |     - |     - |         - |

After:

|                 Method |      N |       Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |------- |-----------:|---------:|---------:|-------:|------:|------:|----------:|
| TupleDescriptor_Create | 100000 | 2,474.8 ns | 12.10 ns | 10.10 ns | 0.1869 |     - |     - |     792 B |
|         Tuple_GetValue | 100000 |   444.2 ns |  2.16 ns |  1.80 ns |      - |     - |     - |         - |
```